pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Felipe Armoni <felipe.armoni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.12,>=3.8"
inference = "^0.15.0"
inference-cli = "^0.15.0"
inference-sdk = "^0.15.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
torch = "^2.3.1"
transformers = "^4.42.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
